% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{normalizeInputMatrix}
\alias{normalizeInputMatrix}
\title{Normalize the input Matrix}
\usage{
normalizeInputMatrix(x)
}
\arguments{
\item{x}{A numeric matrix}
}
\description{
Normalize the input Matrix
}

