% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulateLANDD}
\alias{simulateLANDD}
\title{Simulate LANDD}
\usage{
simulateLANDD(rho, n.sample, z.percent, k, kernel.sd, normalize.method)
}
\arguments{
\item{rho}{Strength of LA correlation}

\item{n.sample}{sample number for LA}

\item{z.percent}{percentage of z in sample}

\item{k}{k step neighbour}

\item{kernel.sd}{standard deviation for kernel}

\item{normalize.method}{Different ways to normalize the result}
}
\value{
simulated data for LANDD
}
\description{
Simulate LANDD
}

